+++ spec/whois/record/parser/responses/whois.nic.la/status_available_spec.rb
@@ -13,7 +13,7 @@
 require 'spec_helper'
 require 'whois/record/parser/whois.nic.la.rb'
 
-describe Whois::Record::Parser::WhoisNicLa, "status_available.expected" do
+describe Whois::Record::Parser::WhoisCentralnicCom, "status_available.expected" do
 
   before(:each) do
     file = fixture("responses", "whois.nic.la/status_available.txt")
@@ -21,6 +21,21 @@ describe Whois::Record::Parser::WhoisNicLa, "status_available.expected" do
     @parser = klass.new(part)
   end
 
+  describe "#disclaimer" do
+    it do
+      @parser.disclaimer.should == nil
+    end
+  end
+  describe "#domain" do
+    it do
+      @parser.domain.should == nil
+    end
+  end
+  describe "#domain_id" do
+    it do
+      @parser.domain_id.should == nil
+    end
+  end
   describe "#referral_whois" do
     it do
       lambda { @parser.referral_whois }.should raise_error(Whois::PropertyNotSupported)
@@ -59,6 +74,29 @@ describe Whois::Record::Parser::WhoisNicLa, "status_available.expected" do
   describe "#expires_on" do
     it do
       @parser.expires_on.should == nil
+    end
+  end
+  describe "#registrar" do
+    it do
+      @parser.registrar.should == nil
+    end
+  end
+  describe "#registrant_contacts" do
+    it do
+      @parser.registrant_contacts.should be_a(Array)
+      @parser.registrant_contacts.should == []
+    end
+  end
+  describe "#admin_contacts" do
+    it do
+      @parser.admin_contacts.should be_a(Array)
+      @parser.admin_contacts.should == []
+    end
+  end
+  describe "#technical_contacts" do
+    it do
+      @parser.technical_contacts.should be_a(Array)
+      @parser.technical_contacts.should == []
     end
   end
   describe "#nameservers" do