+++ spec/whois/record/parser/responses/whois.nic.la/status_registered_spec.rb
@@ -13,7 +13,7 @@
 require 'spec_helper'
 require 'whois/record/parser/whois.nic.la.rb'
 
-describe Whois::Record::Parser::WhoisNicLa, "status_registered.expected" do
+describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected" do
 
   before(:each) do
     file = fixture("responses", "whois.nic.la/status_registered.txt")
@@ -21,6 +21,21 @@ describe Whois::Record::Parser::WhoisNicLa, "status_registered.expected" do
     @parser = klass.new(part)
   end
 
+  describe "#disclaimer" do
+    it do
+      @parser.disclaimer.should == "This whois service is provided by CentralNic Ltd and only contains information pertaining to Internet domain names we have registered for our customers. By using this service you are agreeing (1) not to use any information presented here for any purpose other than determining ownership of domain names (2) not to store or reproduce this data in any way. CentralNic Ltd - www.centralnic.com"
+    end
+  end
+  describe "#domain" do
+    it do
+      @parser.domain.should == "google.la"
+    end
+  end
+  describe "#domain_id" do
+    it do
+      @parser.domain_id.should == "CNIC-DO471480"
+    end
+  end
   describe "#referral_whois" do
     it do
       lambda { @parser.referral_whois }.should raise_error(Whois::PropertyNotSupported)
@@ -62,6 +77,75 @@ describe Whois::Record::Parser::WhoisNicLa, "status_registered.expected" do
     it do
       @parser.expires_on.should be_a(Time)
       @parser.expires_on.should == Time.parse("2010-07-18 23:59:59 UTC")
+    end
+  end
+  describe "#registrar" do
+    it do
+      @parser.registrar.should be_a(_registrar)
+      @parser.registrar.id.should           == "7279-LN"
+      @parser.registrar.name.should         == nil
+      @parser.registrar.organization.should == "LA Registry Pte Ltd"
+      @parser.registrar.url.should          == "http://www.la/"
+    end
+  end
+  describe "#registrant_contacts" do
+    it do
+      @parser.registrant_contacts.should be_a(Array)
+      @parser.registrant_contacts.should have(1).items
+      @parser.registrant_contacts[0].should be_a(_contact)
+      @parser.registrant_contacts[0].type.should          == Whois::Record::Contact::TYPE_REGISTRANT
+      @parser.registrant_contacts[0].id.should            == "ndn-96955"
+      @parser.registrant_contacts[0].name.should          == "Google Inc"
+      @parser.registrant_contacts[0].organization.should  == "Google Inc"
+      @parser.registrant_contacts[0].address.should       == "1600 Amphitheatre Parkway"
+      @parser.registrant_contacts[0].city.should          == "Mountain View"
+      @parser.registrant_contacts[0].zip.should           == "94043"
+      @parser.registrant_contacts[0].state.should         == "Ca"
+      @parser.registrant_contacts[0].country.should       == nil
+      @parser.registrant_contacts[0].country_code.should  == "US"
+      @parser.registrant_contacts[0].phone.should         == "+1.6502530000"
+      @parser.registrant_contacts[0].fax.should           == "+1.65067188571"
+      @parser.registrant_contacts[0].email.should         == "dns-admin@google.com"
+    end
+  end
+  describe "#admin_contacts" do
+    it do
+      @parser.admin_contacts.should be_a(Array)
+      @parser.admin_contacts.should have(1).items
+      @parser.admin_contacts[0].should be_a(_contact)
+      @parser.admin_contacts[0].type.should          == Whois::Record::Contact::TYPE_ADMIN
+      @parser.admin_contacts[0].id.should            == "ndn-96955"
+      @parser.admin_contacts[0].name.should          == "Google Inc"
+      @parser.admin_contacts[0].organization.should  == "Google Inc"
+      @parser.admin_contacts[0].address.should       == "1600 Amphitheatre Parkway"
+      @parser.admin_contacts[0].city.should          == "Mountain View"
+      @parser.admin_contacts[0].zip.should           == "94043"
+      @parser.admin_contacts[0].state.should         == "Ca"
+      @parser.admin_contacts[0].country.should       == nil
+      @parser.admin_contacts[0].country_code.should  == "US"
+      @parser.admin_contacts[0].phone.should         == "+1.6502530000"
+      @parser.admin_contacts[0].fax.should           == "+1.65067188571"
+      @parser.admin_contacts[0].email.should         == "dns-admin@google.com"
+    end
+  end
+  describe "#technical_contacts" do
+    it do
+      @parser.technical_contacts.should be_a(Array)
+      @parser.technical_contacts.should have(1).items
+      @parser.technical_contacts[0].should be_a(_contact)
+      @parser.technical_contacts[0].type.should          == Whois::Record::Contact::TYPE_TECHNICAL
+      @parser.technical_contacts[0].id.should            == "ndn-96955"
+      @parser.technical_contacts[0].name.should          == "Google Inc"
+      @parser.technical_contacts[0].organization.should  == "Google Inc"
+      @parser.technical_contacts[0].address.should       == "1600 Amphitheatre Parkway"
+      @parser.technical_contacts[0].city.should          == "Mountain View"
+      @parser.technical_contacts[0].zip.should           == "94043"
+      @parser.technical_contacts[0].state.should         == "Ca"
+      @parser.technical_contacts[0].country.should       == nil
+      @parser.technical_contacts[0].country_code.should  == "US"
+      @parser.technical_contacts[0].phone.should         == "+1.6502530000"
+      @parser.technical_contacts[0].fax.should           == "+1.65067188571"
+      @parser.technical_contacts[0].email.should         == "dns-admin@google.com"
     end
   end
   describe "#nameservers" do