+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 * NEW: Added full whois.centralnic.com parser.
 
+* NEW: Added full whois.nic.la parser.
+
 * CHANGED: whois.nic.fr `frozen' status is now considered as :registered.
 
 * CHANGED: Rewritten Whois::Record::Parser::Scanners::Base to use a modular approach.